+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicDeploymentManager.scala
@@ -78,7 +78,8 @@ class PeriodicDeploymentManager private[periodic](val delegate: DeploymentManage
                                                   schedulePropertyExtractor: SchedulePropertyExtractor,
                                                   customActionsProvider: PeriodicCustomActionsProvider,
                                                   toClose: () => Unit)
-                                                 (implicit val ec: ExecutionContext) extends DeploymentManager with LazyLogging {
+                                                 (implicit val ec: ExecutionContext)
+  extends DeploymentManager with LazyLogging {
 
 
   override def validate(processVersion: ProcessVersion, deploymentData: DeploymentData, canonicalProcess: CanonicalProcess): Future[Unit] = {
@@ -112,7 +113,8 @@ class PeriodicDeploymentManager private[periodic](val delegate: DeploymentManage
   }
 
   private def cancelIfJobPresent(processVersion: ProcessVersion, user: User): Future[Unit] = {
-    findJobStatus(processVersion.processName)
+    getProcessState(processVersion.processName)(DataFreshnessPolicy.Fresh)
+      .map(_.value)
       .map(_.isDefined)
       .flatMap(shouldStop => {
         if (shouldStop) {
@@ -138,7 +140,7 @@ class PeriodicDeploymentManager private[periodic](val delegate: DeploymentManage
   override def test[T](name: ProcessName, canonicalProcess: CanonicalProcess, scenarioTestData: ScenarioTestData, variableEncoder: Any => T): Future[TestProcess.TestResults[T]] =
     delegate.test(name, canonicalProcess, scenarioTestData, variableEncoder)
 
-  override def findJobStatus(name: ProcessName): Future[Option[ProcessState]] = {
+  override def getProcessState(name: ProcessName)(implicit freshnessPolicy: DataFreshnessPolicy): Future[WithDataFreshnessStatus[Option[ProcessState]]] = {
     def createScheduledProcessState(processDeployment: PeriodicProcessDeployment): ProcessState = {
       processStateDefinitionManager.processState(
         status = ScheduledStatus(processDeployment.runAt),
@@ -196,9 +198,9 @@ class PeriodicDeploymentManager private[periodic](val delegate: DeploymentManage
       errors = state.errors
     )
 
-    delegate
-      .findJobStatus(name)
-      .flatMap {
+    for {
+      delegateState <- delegate.getProcessState(name)
+      postprocessedState <- delegateState.value match {
         // Scheduled again or waiting to be scheduled again.
         case state@Some(js) if js.status.isFinished => handleScheduled(state)
         case state@Some(js) if js.status.isFailed => handleFailed(state)
@@ -207,7 +209,9 @@ class PeriodicDeploymentManager private[periodic](val delegate: DeploymentManage
         // Scheduled or never started or latest run already disappeared in Flink.
         case state@None => handleScheduled(state)
         case Some(js) => Future.successful(Some(js))
-      }.map(_.map(withPeriodicProcessState))
+      }
+      formattedByPeriodicManager = postprocessedState.map(withPeriodicProcessState)
+    } yield WithDataFreshnessStatus(formattedByPeriodicManager, delegateState.cached)
   }
 
   override def processStateDefinitionManager: ProcessStateDefinitionManager =