+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicProcessService.scala
@@ -111,7 +111,7 @@ class PeriodicProcessService(delegateDeploymentManager: DeploymentManager,
 
   //Currently we don't allow simultaneous runs of one scenario - only sequential, so if other schedule kicks in, it'll have to wait
   private def checkIfNotRunning(toDeploy: PeriodicProcessDeployment): Future[Option[PeriodicProcessDeployment]] = {
-    delegateDeploymentManager.findJobStatus(toDeploy.periodicProcess.processVersion.processName).map {
+    delegateDeploymentManager.getProcessState(toDeploy.periodicProcess.processVersion.processName)(DataFreshnessPolicy.Fresh).map(_.value).map {
       case Some(state) if state.isDeployed =>
         logger.debug(s"Deferring run of ${toDeploy.display} as scenario is currently running")
         None
@@ -123,7 +123,7 @@ class PeriodicProcessService(delegateDeploymentManager: DeploymentManager,
 
     def handleSingleProcess(deployedProcess: PeriodicProcessDeployment): Future[Unit] = {
       val processName = deployedProcess.periodicProcess.processVersion.processName
-      delegateDeploymentManager.findJobStatus(processName).flatMap { state =>
+      delegateDeploymentManager.getProcessState(processName)(DataFreshnessPolicy.Fresh).map(_.value).flatMap { state =>
         handleFinishedAction(deployedProcess, state)
           .flatMap { needsReschedule =>
             if (needsReschedule) reschedule(deployedProcess) else scheduledProcessesRepository.monad.pure(()).emptyCallback
@@ -218,10 +218,10 @@ class PeriodicProcessService(delegateDeploymentManager: DeploymentManager,
   }
 
   def deactivate(processName: ProcessName): Future[Unit] = for {
-    status <- delegateDeploymentManager.findJobStatus(processName)
+    status <- delegateDeploymentManager.getProcessState(processName)(DataFreshnessPolicy.Fresh)
     maybePeriodicDeployment <- getLatestDeployment(processName)
     actionResult <- maybePeriodicDeployment match {
-      case Some(periodicDeployment) => handleFinishedAction(periodicDeployment, status)
+      case Some(periodicDeployment) => handleFinishedAction(periodicDeployment, status.value)
         .flatMap(_ => deactivateAction(processName))
         .runWithCallbacks
       case None => deactivateAction(processName).runWithCallbacks