+++ engine/lite/embeddedDeploymentManager/src/test/scala/pl/touk/nussknacker/streaming/embedded/StreamingEmbeddedDeploymentManagerRestartTest.scala
@@ -1,6 +1,6 @@
 package pl.touk.nussknacker.streaming.embedded
 
-import pl.touk.nussknacker.engine.api.deployment.ProcessActionType
+import pl.touk.nussknacker.engine.api.deployment.{DataFreshnessPolicy, ProcessActionType}
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
 import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.engine.build.ScenarioBuilder
@@ -8,6 +8,8 @@ import pl.touk.nussknacker.engine.spel.Implicits._
 
 class StreamingEmbeddedDeploymentManagerRestartTest extends BaseStreamingEmbeddedDeploymentManagerTest {
   import pl.touk.nussknacker.engine.schemedkafka.KafkaUniversalComponentTransformer._
+
+  protected implicit val freshnessPolicy: DataFreshnessPolicy = DataFreshnessPolicy.Fresh
 
   // This test is in separate suite to make sure that restarting of kafka server have no influence on other test case scenarios
   test("Set status to restarting when scenario fails and back to running when the problems are fixed") {
@@ -28,7 +30,7 @@ class StreamingEmbeddedDeploymentManagerRestartTest extends BaseStreamingEmbedde
     kafkaServer.kafkaServer.awaitShutdown()
 
     eventually {
-      val jobStatus = manager.findJobStatus(name).futureValue
+      val jobStatus = manager.getProcessState(name).futureValue.value
       jobStatus.map(_.status) shouldBe Some(SimpleStateStatus.Restarting)
       jobStatus.map(_.allowedActions).get should contain only (ProcessActionType.Cancel)
     }
@@ -36,7 +38,7 @@ class StreamingEmbeddedDeploymentManagerRestartTest extends BaseStreamingEmbedde
     kafkaServer.kafkaServer.startup()
 
     eventually {
-      manager.findJobStatus(name).futureValue.map(_.status) shouldBe Some(SimpleStateStatus.Running)
+      manager.getProcessState(name).futureValue.value.map(_.status) shouldBe Some(SimpleStateStatus.Running)
     }
   }