+++ engine/lite/k8sDeploymentManager/src/test/scala/pl/touk/nussknacker/k8s/manager/BaseK8sDeploymentManagerTest.scala
@@ -93,14 +93,14 @@ class BaseK8sDeploymentManagerTest extends AnyFunSuite with Matchers with Extrem
       } finally {
         manager.cancel(version.processName, DeploymentData.systemUser).futureValue
         eventually {
-          manager.findJobStatus(version.processName).futureValue shouldBe None
+          manager.getFreshProcessState(version.processName).futureValue shouldBe None
         }
       }
     }
 
     def waitForRunning(version: ProcessVersion) = {
       eventually {
-        val state = manager.findJobStatus(version.processName).futureValue
+        val state = manager.getFreshProcessState(version.processName).futureValue
         state.flatMap(_.version) shouldBe Some(version)
         state.map(_.status) shouldBe Some(SimpleStateStatus.Running)
       }