+++ engine/lite/k8sDeploymentManager/src/test/scala/pl/touk/nussknacker/k8s/manager/K8sDeploymentManagerKafkaTest.scala
@@ -71,7 +71,7 @@ class K8sDeploymentManagerKafkaTest extends BaseK8sDeploymentManagerTest
 
     def waitForRunning(version: ProcessVersion) = {
       eventually {
-        val state = manager.findJobStatus(version.processName).futureValue
+        val state = manager.getFreshProcessState(version.processName).futureValue
         state.flatMap(_.version) shouldBe Some(version)
         state.map(_.status) shouldBe Some(SimpleStateStatus.Running)
       }
@@ -279,7 +279,7 @@ class K8sDeploymentManagerKafkaTest extends BaseK8sDeploymentManagerTest
   private def cancelAndAssertCleanup(manager: K8sDeploymentManager, version: ProcessVersion) = {
     manager.cancel(version.processName, DeploymentData.systemUser).futureValue
     eventually {
-      manager.findJobStatus(version.processName).futureValue shouldBe None
+      manager.getFreshProcessState(version.processName).futureValue shouldBe None
     }
     assertNoGarbageLeft()
   }